Added user token functionality and dockerfile

+++ src/main/java/com/spring/catalogapp/controller/ProductController.java
@@ -8,6 +8,7 @@ import com.spring.catalogapp.entity.Product;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,6 +19,8 @@ import java.util.ArrayList;
 @RequestMapping("")
 @RequiredArgsConstructor
 public class ProductController {
+
+    private final String ROLE_ADMIN = "ROLE_EMPLOYEE";
     private final ProductService productService;
     @Autowired
     private final LoginService loginService;
@@ -25,30 +28,78 @@ public class ProductController {
 
     @GetMapping("/products.json")
     public ResponseEntity<ReturnObject> getProducts(@RequestParam int page, @RequestParam int size,
-                                                    @RequestParam String sortElem, @RequestParam String direction) {
+                                                    @RequestParam String sortElem, @RequestParam String direction, @RequestParam String token) {
+
+        if (loginService.getMyUser() == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        if (!loginService.getMyUser().getToken().equals(token)) {
+            return ResponseEntity.notFound().build();
+        }
+        System.out.println(token);
         return new ResponseEntity<>(productService.getProducts(page, size, sortElem, direction), HttpStatus.ACCEPTED);
     }
 
     @PostMapping("/products/save")
-    public ResponseEntity<?> saveProduct(@RequestBody Product product) {
+    public ResponseEntity<?> saveProduct(@RequestBody Product product, @RequestParam String token) {
         productService.extracted(product);
+        if (loginService.getMyUser() == null) {
+            return ResponseEntity.notFound().build();
+        }
 
+        if (!loginService.getMyUser().getToken().equals(token)) {
+            return ResponseEntity.notFound().build();
+        }
+        if (!loginService.getMyUser().getRoles().contains(ROLE_ADMIN)) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok().build();
     }
 
     @PostMapping("/products/filteredList")
-    public ResponseEntity<ArrayList<Product>> filterList(@RequestBody Product product) {
+    public ResponseEntity<ArrayList<Product>> filterList(@RequestBody Product product, @RequestParam String token) {
+        if (loginService.getMyUser() == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        if (!loginService.getMyUser().getToken().equals(token)) {
+            return ResponseEntity.notFound().build();
+        }
+        if (!loginService.getMyUser().getRoles().contains(ROLE_ADMIN)) {
+            return ResponseEntity.notFound().build();
+        }
         return new ResponseEntity<>(productService.filter(product), HttpStatus.ACCEPTED);
     }
 
     @DeleteMapping("/products/delete/{productId}")
-    public ResponseEntity<?> removeProduct(@PathVariable Long productId) {
+    public ResponseEntity<?> removeProduct(@PathVariable Long productId, @RequestParam String token) {
+        if (loginService.getMyUser() == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        if (!loginService.getMyUser().getToken().equals(token)) {
+            return ResponseEntity.notFound().build();
+        }
+        if (!loginService.getMyUser().getRoles().contains(ROLE_ADMIN)) {
+            return ResponseEntity.notFound().build();
+        }
         productService.delete(productId);
         return ResponseEntity.ok().build();
     }
 
     @PutMapping("/products/modify")
-    public ResponseEntity<?> modifyProduct(@RequestBody Product original) {
+    public ResponseEntity<?> modifyProduct(@RequestBody Product original, @RequestParam String token) {
+        if (loginService.getMyUser() == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        if (!loginService.getMyUser().getToken().equals(token)) {
+            return ResponseEntity.notFound().build();
+        }
+        if (!loginService.getMyUser().getRoles().contains(ROLE_ADMIN)) {
+            return ResponseEntity.notFound().build();
+        }
         productService.update(original);
         return ResponseEntity.ok().build();
     }
